Refactor folder to make structure more obvious
