+++ Readme.md
@@ -1,6 +1,18 @@
 # Block Hydration Experiments
 
-A repository to explore possible frontend solutions for Gutenberg.
+A repository to explore patterns related to block hydration with the goal of absorbing as much complexity as possible from the final developers.
+
+_It's not a goal to do an in-depth analysis of the patterns, only to experiment with them in a controlled environment to assess their potential and then test them in [the Gutenberg repository](https://github.com/WordPress/gutenberg), where we will see if they are a good fit or not._
+
+Ideally, the Save component should be derived from the Frontend implementation, and developers should be able to share as much code as possible between the Edit and Save/Frontend component.
+
+There are some folders in this repository:
+
+- `edit`: User code exclusive to the Edit component.
+- `frontend`: User code exclusive to the Frontend component. The Save component is also derived from this.
+- `shared`: User code shared between `edit` and `frontend` components.
+- `gutenberg-packages`: Framwork code that could be absorbed by Gutenberg packages.
+- `webpack`: Bundling configuration that could be absorbed by `@wordpress/scripts` (not existing yet).
 
 Feel free to inspect the code, open issues, submit PRs, ask questions...
 
@@ -16,6 +28,7 @@ Feel free to inspect the code, open issues, submit PRs, ask questions...
 - [ ] Support `useState` and `useEffect` hook in the Save component to be able to reuse the same component in the Frontend: https://github.com/luisherranz/block-hydration-experiments/pull/3.
 - [ ] Support the rest of the hooks in the Save component to be able to reuse the same component in the Frontend.
 - [ ] Support for the Context API between different blocks.
+- [ ] Support for Suspense and Error boundaries between different blocks.
 - [ ] Support for Block Supports (wrapper only):https://github.com/luisherranz/block-hydration-experiments/pull/3.
 - [ ] Avoid the need to populate the `blockProps` (`useBlockProps`).
 - [ ] Support initially hidden Inner blocks.
