+++ src/edit/index.js
@@ -5,7 +5,7 @@
 import "@wordpress/block-editor";
 
 import { useBlockProps, InnerBlocks } from "@wordpress/block-editor";
-import Title from "./frontend/title";
+import Title from "../shared/title";
 
 export default function Edit({ attributes, setAttributes }) {
   const blockProps = useBlockProps();
