+++ src/index.js
@@ -1,11 +1,10 @@
 import "./editor.scss";
 import "./style.scss";
-import { registerBlockType } from "@wordpress/blocks";
+import { registerBlockType } from "./gutenberg-packages/wordpress-blocks";
 import Edit from "./edit";
-import save from "./framework/save";
-import Block from "./frontend/block";
+import Block from "./frontend"; // Save implementation derived from Frontend.
 
 registerBlockType("luisherranz/block-hydration-experiments", {
   edit: Edit,
-  save: save(Block),
+  frontend: Block,
 });
